refactor: extract FlagItem into own class
